Rename properties with expressive names and add sample controller methods

+++ src/main/java/com/jarcelo/henrybooks/models/UserDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
-public interface UserDao extends CrudRepository<User, Double>{
+public interface UserDao extends CrudRepository<User, Long>{
 
 }
